build(simple-ddl): Use vite to build playground
This provides support for reloading the browser when the OCaml source
files change, and provides minification support for the generated
javascript files. It will also make it easier to add javascript
dependencies in the future.

When using the development server, there's sometimes a flash of an error
due to the rebuild racing ahead of dune, meaning that the compiled
javascript file does not exist yet. I’ll get around to fixing that
later.